feat: adiciona endpoints para OrdemServico e ItemOrdemServicoService

+++ Spring/controleServicos/src/main/java/br/edu/ifsp/pep/bcc/controle/demo/service/ItemOrdemServicoService.java
@@ -1,5 +1,6 @@
 package br.edu.ifsp.pep.bcc.controle.demo.service;
 
+import br.edu.ifsp.pep.bcc.controle.demo.controller.exception.NotFoundException;
 import br.edu.ifsp.pep.bcc.controle.demo.model.entities.ItemOrdemServico;
 import br.edu.ifsp.pep.bcc.controle.demo.model.entities.ItemOrdemServicoId;
 import br.edu.ifsp.pep.bcc.controle.demo.repository.ItemOrdemServicoRepository;
@@ -7,7 +8,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -19,18 +19,26 @@ public class ItemOrdemServicoService {
     }
 
     public ItemOrdemServico findById(ItemOrdemServicoId id) {
-        Optional<ItemOrdemServico> optional = itemOrdemServicoRepository.findById(id);
-        if (optional.isPresent()) {
-            return optional.get();
-        }
-        return null;
+        return itemOrdemServicoRepository.findById(id)
+                    .orElseThrow(() -> new NotFoundException("Item não encontrado!"));
     }
 
-    public ItemOrdemServico save(ItemOrdemServico item) {
+    public ItemOrdemServico create(ItemOrdemServico item) {
         return itemOrdemServicoRepository.save(item);
     }
 
+//    public ItemOrdemServico update(ItemOrdemServicoId id, ItemOrdemServico itemAtualizado) {
+//        if (!itemOrdemServicoRepository.existsById(id)) {
+//            throw new NotFoundException("Item não encontrado!");
+//        }
+//        itemAtualizado.setNumeroItem(new ItemOrdemServicoId(id.getOrdemServico(), id.getOrdemServico()));
+//        return itemOrdemServicoRepository.save(itemAtualizado);
+//    }
+
     public void deleteById(ItemOrdemServicoId id) {
+        if (!itemOrdemServicoRepository.existsById(id)) {
+            throw new NotFoundException("Item não encontrado!");
+        }
         itemOrdemServicoRepository.deleteById(id);
     }
 }
